requirements.txt
six
enum34
invoke
requests
decorator
inflection
python-dateutil
